ME - Turnitin upgrade.

+++ mod/turnitintooltwo/turnitintooltwo_submission.class.php
@@ -14,6 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
+require_once(__DIR__.'/classes/digitalreceipt/receipt_message.php');
+
 class turnitintooltwo_submission {
 
     public $id;
@@ -44,8 +46,11 @@ class turnitintooltwo_submission {
     public $submission_orcapable;
     public $submission_acceptnothing;
     public $overall_grade;
+    private $receipt;
 
     public function __construct($id = 0, $idtype = "moodle", $turnitintooltwoassignment = "", $partid = "") {
+        $this->receipt = new receipt_message();
+
         if ($idtype == "turnitin") {
             $this->submission_objectid = $id;
             $this->submission_part = $partid;
@@ -91,11 +96,6 @@ class turnitintooltwo_submission {
         if (!$this->id = $DB->insert_record('turnitintooltwo_submissions', $submission)) {
             return false;
         }
-        $turnitintooltwoassignment = new turnitintooltwo_assignment($this->turnitintooltwoid);
-        $cm = get_coursemodule_from_instance("turnitintooltwo", $turnitintooltwoassignment->turnitintooltwo->id,
-            $turnitintooltwoassignment->turnitintooltwo->course);
-
-        turnitintooltwo_add_to_log($turnitintooltwoassignment->turnitintooltwo->course, "add submission", 'view.php?id='.$cm->id, get_string('addsubmissiondesc', 'turnitintooltwo') . " '" . $data['submissiontitle'] . "'", $cm->id, $data['studentsname']);
 
         $this->reset_submission($data);
 
@@ -148,7 +148,8 @@ class turnitintooltwo_submission {
             $condition = array("submission_objectid" => $this->submission_objectid);
         }
 
-        if ($submission = $DB->get_record('turnitintooltwo_submissions', $condition)) {
+        if ($submission = $DB->get_record('turnitintooltwo_submissions',
+                                            $condition, '*', IGNORE_MULTIPLE)) {
             if (empty($turnitintooltwoassignment)) {
                 $turnitintooltwoassignment = new turnitintooltwo_assignment($submission->turnitintooltwoid);
                 $this->turnitintooltwoid = $turnitintooltwoassignment->turnitintooltwo->id;
@@ -269,7 +270,7 @@ class turnitintooltwo_submission {
      * @param int $submission_id
      * @return type
      */
-    public function delete_submission() {
+    public function delete_submission($action = 'delete') {
         global $CFG, $DB;
         $notice = array();
         $partid = $this->submission_part;
@@ -281,13 +282,25 @@ class turnitintooltwo_submission {
         $cm = get_coursemodule_from_instance("turnitintooltwo", $turnitintooltwoassignment->turnitintooltwo->id,
             $turnitintooltwoassignment->turnitintooltwo->course);
 
-        turnitintooltwo_add_to_log($turnitintooltwoassignment->turnitintooltwo->course, "delete submission", 'view.php?id='.$cm->id, get_string('deletesubmissiondesc', 'turnitintooltwo') . " '$this->submission_title'", $cm->id, $this->userid);
+        $istutor = has_capability('mod/turnitintooltwo:grade', context_module::instance($cm->id));
 
         // Delete Moodle submission first.
         if (!$DB->delete_records('turnitintooltwo_submissions', array('id' => $this->id))) {
             $notice["type"] = "error";
             $notice["message"] = get_string('submissiondeleteerror', 'turnitintooltwo');
             return $notice;
+        }
+
+        // Log deleted submission with Moodle.
+        if ($action == 'delete') {
+            turnitintooltwo_add_to_log(
+                $turnitintooltwoassignment->turnitintooltwo->course,
+                "delete submission",
+                'view.php?id='.$cm->id,
+                get_string('deletesubmissiondesc', 'turnitintooltwo') . " '$this->submission_title'",
+                $cm->id,
+                $this->userid
+            );
         }
 
         // Update grade in Gradecenter.
@@ -311,24 +324,47 @@ class turnitintooltwo_submission {
         $grades->userid = $this->userid;
         $params['idnumber'] = $cm->idnumber;
 
+        // If this is the only submission and anon marking is being used then unlock this part.
+        $numpartsubs = $DB->count_records('turnitintooltwo_submissions',
+                                            array('submission_part' => $this->submission_part));
+
+        if ($numpartsubs == 0) {
+            $unlockpart = new stdClass();
+            $unlockpart->id = $this->submission_part;
+            $unlockpart->unanon = 0;
+            $DB->update_record('turnitintooltwo_parts', $unlockpart);
+        }
+
         @include_once($CFG->libdir."/gradelib.php");
         grade_update('mod/turnitintooltwo', $turnitintooltwoassignment->turnitintooltwo->course, 'mod',
                         'turnitintooltwo', $turnitintooltwoassignment->turnitintooltwo->id, 0, $grades, $params);
 
         // If we have a Turnitin Id then delete submission.
-        if (!empty($this->submission_objectid)) {
+        if ((!empty($this->submission_objectid)) && ($istutor) && $action == 'delete') {
             $submission = new TiiSubmission();
             $submission->setSubmissionId($this->submission_objectid);
 
             try {
                 $response = $turnitincall->deleteSubmission($submission);
 
-                $notice["message"] = get_string('submissiondeleted', 'turnitintooltwo');
+                turnitintooltwo_add_to_log(
+                    $turnitintooltwoassignment->turnitintooltwo->course,
+                    "delete submission",
+                    'view.php?id='.$cm->id,
+                    get_string('deletesubmissiontiidesc', 'turnitintooltwo') . " '$this->submission_title'",
+                    $cm->id,
+                    $this->userid
+                );
+
+                $notice["type"] = "full-error";
+                $notice["message"] = get_string('submissiondeleted', 'turnitintooltwo').
+                                        ' ('.get_string('turnitinid', 'turnitintooltwo').
+                                            ': '.$this->submission_objectid.')';
 
                 // If we have no submissions to this part then reset submitted and unanon flag.
                 $numsubs = count($DB->get_records('turnitintooltwo_submissions',
                                             array('submission_part' => $partid), 'id'));
-                
+
                 return $notice;
             } catch (Exception $e) {
                 $turnitincomms->handle_exceptions($e, 'turnitindeletionerror');
@@ -346,7 +382,7 @@ class turnitintooltwo_submission {
      * @return integer $submissionid the submissions id for the submission
      */
     public function do_tii_nothing_submission($cm, $turnitintooltwoassignment, $partid, $userid) {
-        global $DB, $USER;
+        global $DB, $USER, $CFG;
 
         $context = context_module::instance($cm->id);
 
@@ -354,6 +390,7 @@ class turnitintooltwo_submission {
         $course = $turnitintooltwoassignment->get_course_data($turnitintooltwoassignment->turnitintooltwo->course);
         $user = new turnitintooltwo_user($userid, 'Learner');
         $user->join_user_to_class($course->turnitin_cid);
+        $user->edit_tii_user();
 
         $part = $turnitintooltwoassignment->get_part_details($partid);
 
@@ -388,6 +425,25 @@ class turnitintooltwo_submission {
             $submission->submission_acceptnothing = 1;
             $submission->submission_orcapable = 0;
 
+            //Send a message to the user's Moodle inbox with the digital receipt.
+            $partdetails = $turnitintooltwoassignment->get_part_details($partid);
+            $input = array(
+                'firstname' => $user->firstname,
+                'lastname' => $user->lastname,
+                'submission_title' => get_string('gradingtemplate', 'turnitintooltwo'),
+                'assignment_name' => $turnitintooltwoassignment->turnitintooltwo->name,
+                'assignment_part' => $partdetails->partname,
+                'course_fullname' => $course->fullname,
+                'submission_date' => date('d-M-Y h:iA'),
+                'submission_id' => $submission->submission_objectid
+            );
+
+            $message = $this->receipt->build_message($input);
+            $this->receipt->send_message($userid, $message);
+
+            //Add entry to log.
+            turnitintooltwo_add_to_log($turnitintooltwoassignment->turnitintooltwo->course, "add submission", 'view.php?id='.$cm->id, get_string('gradenosubmission', 'turnitintooltwo') . ": $userid", $cm->id, $userid);
+
             if (!$this->id = $DB->insert_record('turnitintooltwo_submissions', $submission)) {
                 return get_string('submissionupdateerror', 'turnitintooltwo');
             } else {
@@ -396,10 +452,16 @@ class turnitintooltwo_submission {
                 $assignment->submitted = 1;
                 $DB->update_record('turnitintooltwo', $assignment);
 
-                $part = new stdClass();
-                $part->id = $partid;
-                $part->submitted = 1;
-                $DB->update_record('turnitintooltwo_parts', $part);
+                $part_data = new stdClass();
+                $part_data->id = $partid;
+                $part_data->submitted = 1;
+
+                //Disable anonymous marking if post date has passed.
+                if ($part->dtpost <= time()) {
+                    $part_data->unanon = 1;
+                }
+
+                $DB->update_record('turnitintooltwo_parts', $part_data);
 
                 return array( "submission_id" => $newsubmission->getSubmissionId() );
             }
@@ -407,6 +469,7 @@ class turnitintooltwo_submission {
         } catch (Exception $e) {
             return $e->getMessage();
         }
+
         return $data;
     }
 
@@ -418,15 +481,17 @@ class turnitintooltwo_submission {
      * @return string $message to display to user
      */
     public function do_tii_submission($cm, $turnitintooltwoassignment) {
-        global $DB, $USER;
+        global $DB, $USER, $CFG;
 
-        $notice = array();
+        $config = turnitintooltwo_admin_config();
+        $notice = array("success" => false);
         $context = context_module::instance($cm->id);
 
         // Check if user is a member of class, if not then join them to it.
         $course = $turnitintooltwoassignment->get_course_data($turnitintooltwoassignment->turnitintooltwo->course);
         $user = new turnitintooltwo_user($this->userid, 'Learner');
         $user->join_user_to_class($course->turnitin_cid);
+        $user->edit_tii_user();
 
         // Get the stored file and read it into a temp file for submitting to Turnitin.
         $fs = get_file_storage();
@@ -440,7 +505,7 @@ class turnitintooltwo_submission {
                 $cm->id
             );
 
-            if ( ! $turnitintooltwoassignment->turnitintooltwo->anon) {
+            if ( !$turnitintooltwoassignment->turnitintooltwo->anon && empty($config->enablepseudo) ) {
                 $user_details = array(
                     $this->userid,
                     $user->firstname,
@@ -531,12 +596,34 @@ class turnitintooltwo_submission {
                 $notice["extract"] = $newsubmission->getTextExtract();
                 $notice["tii_submission_id"] = $submission->submission_objectid;
 
+                //Send a message to the user's Moodle inbox with the digital receipt.
+                $input = array(
+                    'firstname' => $user->firstname,
+                    'lastname' => $user->lastname,
+                    'submission_title' => $this->submission_title,
+                    'assignment_name' => $turnitintooltwoassignment->turnitintooltwo->name,
+                    'assignment_part' => $partdetails = $turnitintooltwoassignment->get_part_details($this->submission_part)->partname,
+                    'course_fullname' => $course->fullname,
+                    'submission_date' => date('d-M-Y h:iA'),
+                    'submission_id' => $submission->submission_objectid
+                );
+
+                $message = $this->receipt->build_message($input);
+                $this->receipt->send_message($this->userid, $message);
+
+                //Create a log entry for submission going to Turnitin.
+                $logstring = ($apimethod == "replaceSubmission") ? 'addresubmissiontiidesc' : 'addsubmissiontiidesc';
+
+                turnitintooltwo_add_to_log(
+                    $turnitintooltwoassignment->turnitintooltwo->course,
+                    "add submission",
+                    'view.php?id='.$cm->id,
+                    get_string($logstring, 'turnitintooltwo') . " '" . $this->submission_title . "'",
+                    $cm->id,
+                    $user->id
+                );
             } catch (Exception $e) {
-                if (!is_null($this->submission_objectid)) {
-                    $errorstring = "updatesubmissionerror";
-                } else {
-                    $errorstring = "createsubmissionerror";
-                }
+                $errorstring = (!is_null($this->submission_objectid)) ? "updatesubmissionerror" : "createsubmissionerror";
                 $error = $turnitincomms->handle_exceptions($e, $errorstring, false, true);
 
                 $notice["message"] = $error;
@@ -587,7 +674,8 @@ class turnitintooltwo_submission {
         global $DB, $CFG;
 
         static $part;
-        if (empty($part)) {
+        static $tiiassignid;
+        if ($tiiassignid != $tiisubmissiondata->getAssignmentId() || empty($part)) {
             $part = $DB->get_record("turnitintooltwo_parts", array("tiiassignid" => $tiisubmissiondata->getAssignmentId()));
         }
         $turnitintooltwoassignment = new turnitintooltwo_assignment($part->turnitintooltwoid);
@@ -633,6 +721,7 @@ class turnitintooltwo_submission {
         if ($save) {
             // If the user is not a moodle user then get their name from Tii - only do this on initial save.
             $sub->userid = turnitintooltwo_user::get_moodle_user_id($tiisubmissiondata->getAuthorUserId());
+
             if ($sub->userid == 0 && empty($this->id)) {
                 if ($tiisubmissiondata->getAuthorUserId() > 0) {
                     $sub->submission_nmuserid = $tiisubmissiondata->getAuthorUserId();
